fixup! Refactor of the links file.
