+++ libraries/rush-lib/src/utilities/actionNameConstants.ts
@@ -1,5 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
+export const PURGE_ACTION_NAME: 'purge' = 'purge';
 export const LINK_PACKAGE_ACTION_NAME: 'link-package' = 'link-package';
 export const BRIDGE_PACKAGE_ACTION_NAME: 'bridge-package' = 'bridge-package';
